+++ CHANGELOG.md
@@ -39,6 +39,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   versioned feature policy in
   [#857](https://github.com/bitshifter/glam-rs/issues/857).
 
+- Added a `speedy-08` feature that adds the `speedy` 0.8 impls, following the
+  versioned feature policy in
+  [#857](https://github.com/bitshifter/glam-rs/issues/857).
+
 ### Changed
 
 - `FloatExt::lerp` for `f32` and `f64` now uses the same form as the vector `lerp` methods,
@@ -56,6 +60,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   major release.
 
 - The `mint` feature, which is an alias for `mint-05` and will be removed in the next
+  major release.
+
+- The `speedy` feature, which is an alias for `speedy-08` and will be removed in the next
   major release.
 
 ### Fixed